requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
scikit-learn
pandas
python-dotenv
gunicorn